backend/requirements.txt
boto3
yfinance
strands-agents
strands-agents-tools
fredapi
python-dotenv
newsapi-python